refactor: Move build environment exceptions to their own submodule
